pattern-factory/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Security
PyJWT>=2.8.0
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4

# Scalability
redis>=5.0.0

# Geometric operations
shapely>=2.0.0
numpy>=1.24.0

# Database
supabase>=2.0.0
postgrest>=0.13.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# CLI
click>=8.1.0
rich>=13.0.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional - for advanced nesting
scipy>=1.11.0
matplotlib>=3.7.0
pillow>=10.0.0
